Refactor CompraEntity to use double for precio; add method for random decimal generation in AleatorioService; update fill methods in CompraService and FacturaService to utilize new random pricing; enhance TemporadaService with dynamic naming for seasons.

+++ src/main/java/net/ausiasmarch/gesportin/entity/CompraEntity.java
@@ -33,7 +33,7 @@ public class CompraEntity {
 
     @NotNull
     @Column(nullable = false)
-    private BigDecimal precio;
+    private double precio;
 
     @NotNull
     @ManyToOne(fetch = FetchType.EAGER)
